requirements.txt
selene == 2.0.0b17
pytest
allure-pytest
python-dotenv
openpyxl
